CLOUDP-290416: Testing for ipa s3 dataDump
