+++ package.json
@@ -35,6 +35,7 @@
         "@eslint/js": "^9.18.0",
         "@jest/globals": "^29.7.0",
         "@stoplight/types": "^14.1.1",
+        "aws-sdk-client-mock": "^4.1.0",
         "babel-plugin-transform-import-meta": "^2.3.2",
         "eslint": "^9.18.0",
         "eslint-plugin-require-extensions": "^0.1.3",
